test: use Jest
