+++ package.json
@@ -10,8 +10,8 @@
   "scripts": {
     "clean": "tsc --build --clean",
     "build": "tsc --build --verbose --listEmittedFiles",
-    "test:unit": "tape test/*.js",
-    "test": "pnpm run build && standard && pnpm run test:unit && mos test",
+    "test:unit": "jest",
+    "test": "pnpm run build && standard --env jest && pnpm run test:unit && mos test",
     "md": "mos",
     "prepublishOnly": "pnpm run build"
   },
@@ -33,14 +33,13 @@
   "devDependencies": {
     "@types/is-windows": "^0.2.0",
     "@types/node": "^12.7.2",
+    "jest": "^24.9.0",
     "lodash": "^4.17.15",
     "memory-fs": "^0.4.1",
     "mock-fs": "^4.10.1",
     "mos": "^1.3.1",
     "mos-plugin-readme": "^1.0.4",
     "standard": "^14.0.2",
-    "tape": "^4.11.0",
-    "tape-promise": "^4.0.0",
     "typescript": "^3.5.3"
   },
   "engines": {
@@ -50,5 +49,8 @@
     "plugins": [
       "readme"
     ]
+  },
+  "jest": {
+    "testEnvironment": "node"
   }
 }
